test(e2e): address pass-4 bot review (1 P1 + 1 P2)
- error-recovery: split file in two so the protected (/works, /en bell)
  cases run under the authenticated chromium project and the login-401
  case runs unauth. Codex P1: under chromium-no-auth, the dashboard
  layout returns null when getAuthFromCookie() has no user, so the
  /works recovery test was observing an empty <body> before the route
  mock could fire. Reverted the playwright.config.ts routing for
  error-recovery; new error-recovery-unauth.spec.ts takes the no-auth
  slot and also adds a 500-response twin.
- audit-log-immutable: extractor now reads `body.activities` (the actual
  controller response shape — apps/api/src/activity-log/activity-log.controller.ts
  returns `{ activities, total }`). Kept `entries`/`data`/`logs` as
  fallbacks for robustness. Codex P2: against the current API the tests
  were skipping with "no entries" after creating a work, so PATCH/PUT/
  DELETE immutability regressions wouldn't have been caught.
