+++ apps/web/playwright.config.ts
@@ -43,7 +43,7 @@ export default defineConfig({
             // they assert things like "/works redirects to login" which is
             // only true when the user is signed out.
             testIgnore:
-                /\/(auth|navigation|password-reset|user-journey|works-public|works-api|api-public-contract|notifications|accessibility|seo-meta|error-pages|website-templates|subscriptions|conversations|git-providers|forms-validation|i18n-locales|screenshot-and-deploy|health-meta|performance-budget|responsive-viewports|error-recovery|keyboard-navigation)\.spec\.ts$/,
+                /\/(auth|navigation|password-reset|user-journey|works-public|works-api|api-public-contract|notifications|accessibility|seo-meta|error-pages|website-templates|subscriptions|conversations|git-providers|forms-validation|i18n-locales|screenshot-and-deploy|health-meta|performance-budget|responsive-viewports|error-recovery-unauth|keyboard-navigation)\.spec\.ts$/,
             use: {
                 ...devices['Desktop Chrome'],
                 storageState: './e2e/.auth/user.json',
@@ -54,7 +54,7 @@ export default defineConfig({
         {
             name: 'chromium-no-auth',
             testMatch:
-                /\/(auth|navigation|password-reset|user-journey|works-public|works-api|api-public-contract|notifications|accessibility|seo-meta|error-pages|website-templates|subscriptions|conversations|git-providers|forms-validation|i18n-locales|screenshot-and-deploy|health-meta|performance-budget|responsive-viewports|error-recovery|keyboard-navigation)\.spec\.ts$/,
+                /\/(auth|navigation|password-reset|user-journey|works-public|works-api|api-public-contract|notifications|accessibility|seo-meta|error-pages|website-templates|subscriptions|conversations|git-providers|forms-validation|i18n-locales|screenshot-and-deploy|health-meta|performance-budget|responsive-viewports|error-recovery-unauth|keyboard-navigation)\.spec\.ts$/,
             use: {
                 ...devices['Desktop Chrome'],
             },
